+++ .eslintrc.cjs
@@ -18,4 +18,5 @@ module.exports = {
       { allowConstantExport: true },
     ],
   },
+  'react/prop-types': 'off',
 }
