+++ package.json
@@ -11,6 +11,7 @@
     "clear-vite-cache": "rm -rf .vite"
   },
   "dependencies": {
+    "axios": "^1.6.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.2.2"
